safety_ai_app/pyproject.toml
# pyproject.toml

[tool.poetry]
name = "safety-ai-app"
version = "0.1.0"
description = "Safety AI App"
authors = ["Leo <leo@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "safety_ai_app", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"

# Core dependencies
openai = ">=2.26.0"
streamlit = ">=1.56.0"
sentence-transformers = "^5.1.0"
pypdf = ">=6.10.0"
markdown = "^3.9"
google-auth-oauthlib = "^1.2.2"
google-api-python-client = "^2.181.0"
google-auth-httplib2 = "^0.2.0"
requests = ">=2.33.0"

# Security-pinned transitive dependencies (minimum patched versions)
urllib3 = ">=2.6.3"
pyasn1 = ">=0.6.3"
tornado = ">=6.5.5"
protobuf = ">=6.33.5,<7.0.0"
langsmith = ">=0.7.31"
aiohttp = ">=3.11.14"

# ChromaDB
chromadb = "^1.1.0"

# Langchain ecosystem - pinned to compatible versions (security-patched)
langchain-core = "^1.0.2"
langchain-openai = ">=1.1.14"
langchain-text-splitters = ">=1.1.2"
langchain-classic = "^1.0.0"
langchain = ">=1.0.0"
langgraph-checkpoint = ">=4.0.0"
transformers = ">=5.0.0"

# Para o timer do quiz
streamlit-autorefresh = "^1.0.1"

# Outras dependencias
python-docx = "^1.2.0"
python-dotenv = "^1.1.1"
rank-bm25 = "^0.2.2"
ddgs = "^9.6.0"
feedparser = "^6.0.10"
pytz = "^2024.1"

pymupdf = "^1.26.4"
pytesseract = "^0.3.13"
pdf2image = "^1.17.0"
pillow = ">=12.2.0"
python-pptx = "^1.0.2"
pandas = "^2.3.3"
openpyxl = "^3.1.5"
httpx = "^0.28.1"
certifi = "^2025.10.5"
httplib2 = "^0.31.0"
streamlit-drawable-canvas = "^0.9.3"
docx2pdf = "^0.1.8"
docxtpl = "^0.20.1"
beautifulsoup4 = "^4.14.2"
pypdf2 = "^3.0.1"
langchain-chroma = "^1.0.0"
langchain-community = "^0.4.1"
pymdown-extensions = "^10.16.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3"
ruff = "^0.5"
black = "^24.8"
mypy = "^1.11"
pytest-cov = "^5.0"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
